requirements.txt
# Core dependencies
torch==2.3.0
torchvision>=0.13.0
numpy>=1.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
tqdm>=4.60.0
Pillow>=8.0.0
PyYAML>=5.4.0

# Optional dependencies
efficientnet-pytorch>=0.7.0

# Development dependencies (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
tensorboard>=2.7.0
